+++ CHANGELOG.md
@@ -7,7 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
-<!-- ### Added -->
+### Added
+
+- feat(bot)!: interaction server support
 
 ### Changed